If --debug was used, don't delete the temporary .perl file used to pass options to latex2html.

+++ Doc/tools/mkhowto.sh
@@ -376,4 +376,6 @@ for FILE in $@ ; do
     rm -f $LOGFILE
 done
 
-rm -f $L2H_AUX_INIT_FILE
+if [ ! "$DEBUGGING" ] ; then
+    rm -f $L2H_AUX_INIT_FILE
+fi
